`ImmutableBase` -> `Chain`, and minor changes
