+++ tests/index.js
@@ -1,10 +1,3 @@
-import 'babel-polyfill';
-
-import SouceMapSupport from 'source-map-support';
-SouceMapSupport.install();
-
-import 'tape-catch';
-import './custom-tests';
-
+import './test';
 import './unit';
 import './integration';
