+++ tests/test.js
@@ -1,7 +1,8 @@
+import 'babel-polyfill';
+
 import SouceMapSupport from 'source-map-support';
 SouceMapSupport.install();
 
 import 'tape-catch';
 import './custom-tests';
-import test from 'tape';
-export default test;
+export { default as default } from 'tape';
