+++ tests/unit/index.js
@@ -1,6 +1,6 @@
 import './arguments';
 import './atlas';
-import './immutable-base';
+import './chain';
 import './mapper';
 import './registry';
 import './related';
