+++ tests/unit/mapper/insert.js
@@ -6,333 +6,328 @@ import CamelCase from '../../../lib/plugins/CamelCase';
 
 const Pg = Mapper.knex(Knex({ client: 'pg' }));
 
-test('== Mapper - insert ==', t => {
+test('Mapper.insert()', st => {
 
-  t.test('Mapper.insert()', t => {
+  st.plan(5);
 
-    t.plan(5);
+  st.throws(
+    () => Mapper.insert(),
+    TypeError,
+    'throws `TypeError` synchronously with no arguments'
+  );
 
-    t.throws(
-      () => Mapper.insert(),
-      TypeError,
-      'throws `TypeError` synchronously with no arguments'
-    );
+  st.resolvesTo(
+    Mapper.insert(null), null,
+    'resolves `null` to `null`'
+  );
 
-    t.resolvesTo(
-      Mapper.insert(null), null,
-      'resolves `null` to `null`'
-    );
+  st.resolvesToDeep(
+    Mapper.insert([]), [],
+    'resolves empty array to empty array'
+  );
 
-    t.resolvesToDeep(
-      Mapper.insert([]), [],
-      'resolves empty array to empty array'
-    );
+  st.resolvesToDeep(
+    Mapper.insert([null, null]), [],
+    'resolves array of `null` values to an empty array'
+  );
 
-    t.resolvesToDeep(
-      Mapper.insert([null, null]), [],
-      'resolves array of `null` values to an empty array'
-    );
+  st.resolvesToDeep(
+    Mapper.insert(null, null), [],
+    'resolves multiple `null` value arguments to an empty array'
+  );
+});
 
-    t.resolvesToDeep(
-      Mapper.insert(null, null), [],
-      'resolves multiple `null` value arguments to an empty array'
-    );
-  });
+test('Mapper.prepareInsert()', st => {
 
-  t.test('Mapper.prepareInsert()', t => {
+  const Records = Mapper.table('records').prepareInsert(
+    { text: 'a' }
+  );
 
-    const Records = Mapper.table('records').prepareInsert(
-      { text: 'a' }
-    );
+  st.queriesEqual(
+    Records.toQueryBuilder(),
+    `insert into "records" ("text") values ('a')`,
+    `single record - SQL`
+  );
 
-    t.queriesEqual(
-      Records.toQueryBuilder(),
-      `insert into "records" ("text") values ('a')`,
-      `single record - SQL`
-    );
+  const Things = Pg.table('things').prepareInsert(
+    { item: 'a' }
+  );
 
-    const Things = Pg.table('things').prepareInsert(
-      { item: 'a' }
-    );
+  st.queriesEqual(
+    Things.toQueryBuilder(),
+    `insert into "things" ("item") values ('a') returning *`,
+    `single record - PostgreSQL`
+  );
 
-    t.queriesEqual(
-      Things.toQueryBuilder(),
-      `insert into "things" ("item") values ('a') returning *`,
-      `single record - PostgreSQL`
-    );
+  const Apples = Mapper.table('apples').prepareInsert([
+    { color: 'red' },
+    { color: 'green' }
+  ]);
 
-    const Apples = Mapper.table('apples').prepareInsert([
-      { color: 'red' },
-      { color: 'green' }
-    ]);
+  st.queriesEqual(
+    Apples.toQueryBuilder(),
+    `insert into "apples" ("color") values ('red'), ('green')`,
+    `multiple records`
+  );
 
-    t.queriesEqual(
-      Apples.toQueryBuilder(),
-      `insert into "apples" ("color") values ('red'), ('green')`,
-      `multiple records`
-    );
+  st.end();
+});
 
-    t.end();
-  });
+test('Mapper.handleInsertOneResponse()', st => {
 
-  t.test('Mapper.handleInsertOneResponse()', t => {
+  const Records = Mapper.idAttribute('record_id');
 
-    const Records = Mapper.idAttribute('record_id');
+  st.deepEqual(
+    Records.handleInsertOneResponse([5], { text: 'a' }),
+    { record_id: 5, text: 'a' },
+    'ID array response'
+  );
 
-    t.deepEqual(
-      Records.handleInsertOneResponse([5], { text: 'a' }),
-      { record_id: 5, text: 'a' },
-      'ID array response'
-    );
+  st.deepEqual(
+    Records.handleInsertOneResponse([], { item: 'a' }),
+    { item: 'a' },
+    'Empty response'
+  );
 
-    t.deepEqual(
-      Records.handleInsertOneResponse([], { item: 'a' }),
-      { item: 'a' },
-      'Empty response'
-    );
+  st.end();
+});
 
-    t.end();
-  });
+test(
+  'Mapper.handleInsertOneResponse() - respects `columnToAttribute`'
+, st => {
 
-  t.test(
-    'Mapper.handleInsertOneResponse() - respects `columnToAttribute`'
-  , st => {
+  const Records = Mapper.extend(CamelCase()).idAttribute('recordId');
 
-    const Records = Mapper.extend(CamelCase()).idAttribute('recordId');
+  st.deepEqual(
+    Records.handleInsertOneResponse([5], { someValue: 'a' }),
+    { recordId: 5, someValue: 'a' },
+    'ID array response'
+  );
 
-    st.deepEqual(
-      Records.handleInsertOneResponse([5], { someValue: 'a' }),
-      { recordId: 5, someValue: 'a' },
-      'ID array response'
-    );
+  const objectResponse = [{ record_id: 5, some_value: 'a' }];
+  st.deepEqual(
+    Records.handleInsertOneResponse(objectResponse, { someValue: 'a' }),
+    { recordId: 5, someValue: 'a' },
+    'ID array response'
+  );
 
-    const objectResponse = [{ record_id: 5, some_value: 'a' }];
-    st.deepEqual(
-      Records.handleInsertOneResponse(objectResponse, { someValue: 'a' }),
-      { recordId: 5, someValue: 'a' },
-      'ID array response'
-    );
-
-    st.end();
-  });
+  st.end();
+});
 
 
-  t.test('Mapper.handleInsertManyResponse() - single ID response', t => {
+test('Mapper.handleInsertManyResponse() - single ID response', st => {
 
-    const records = [
-      { code: 1234, color: 'red' },
-      { color: 'green' }
-    ];
+  const records = [
+    { code: 1234, color: 'red' },
+    { color: 'green' }
+  ];
 
-    const Records = Mapper.idAttribute('code');
+  const Records = Mapper.idAttribute('code');
 
-    t.deepEqual(
-      Records.handleInsertManyResponse([1234], records),
-      [{ code: 1234, color: 'red' }, { color: 'green' }],
-      'assigns ID attribute to first record only'
-    );
+  st.deepEqual(
+    Records.handleInsertManyResponse([1234], records),
+    [{ code: 1234, color: 'red' }, { color: 'green' }],
+    'assigns ID attribute to first record only'
+  );
 
-    t.end();
-  });
+  st.end();
+});
 
-  t.test('Mapper.handleInsertManyResponse() - PostgreSQL response', t => {
-    const Oranges = Pg.table('oranges');
-    const oranges = [{ name: 'valencia' }, { name: 'navel'}];
+test('Mapper.handleInsertManyResponse() - PostgreSQL response', st => {
+  const Oranges = Pg.table('oranges');
+  const oranges = [{ name: 'valencia' }, { name: 'navel'}];
 
-    const response = [
+  const response = [
+    { id: 1, name: 'valencia', created_at: 'some_date' },
+    { id: 2, name: 'navel', created_at: 'some_time' }
+  ];
+
+  const result = Oranges.handleInsertManyResponse(response, oranges);
+
+  st.deepEqual(oranges[0], { name: 'valencia' }, 'record 0 is unchanged');
+  st.deepEqual(oranges[1], { name: 'navel' }, 'record 1 is unchanged');
+
+  st.deepEqual(
+    result,
+    [
       { id: 1, name: 'valencia', created_at: 'some_date' },
       { id: 2, name: 'navel', created_at: 'some_time' }
-    ];
+    ],
+    'returns updates records'
+  );
 
-    const result = Oranges.handleInsertManyResponse(response, oranges);
+  st.end();
+});
 
-    t.deepEqual(oranges[0], { name: 'valencia' }, 'record 0 is unchanged');
-    t.deepEqual(oranges[1], { name: 'navel' }, 'record 1 is unchanged');
+test(
+  'Mapper.handleInsertManyResponse() - object response - ' +
+  'respects `attributeToColumn`',
+st => {
 
-    t.deepEqual(
-      result,
-      [
-        { id: 1, name: 'valencia', created_at: 'some_date' },
-        { id: 2, name: 'navel', created_at: 'some_time' }
-      ],
-      'returns updates records'
-    );
+  const Oranges = Mapper.extend(CamelCase())
+    .table('oranges').idAttribute('orangeId');
 
-    t.end();
+  const oranges = [{ orangeType: 'valencia' }, { orangeType: 'navel'}];
+
+  const response = [
+    { orange_id: 1, orange_type: 'valencia', created_at: 'some_date' },
+    { orange_id: 2, orange_type: 'navel', created_at: 'some_time' }
+  ];
+
+  st.deepEqual(
+    Oranges.handleInsertManyResponse(response, oranges),
+    [
+      { orangeId: 1, orangeType: 'valencia', createdAt: 'some_date' },
+      { orangeId: 2, orangeType: 'navel', createdAt: 'some_time' }
+    ],
+    'returns updated records'
+  );
+
+  st.end();
+});
+
+
+test(
+  'Mapper.handleInsertManyResponse() - idResponse - ' +
+  'respects `attributeToColumn`',
+st => {
+
+  const Oranges = Mapper.extend(CamelCase())
+    .table('oranges').idAttribute('orangeId');
+
+  const oranges = [{ orangeType: 'valencia' }, { orangeType: 'navel'}];
+
+  st.deepEqual(
+    Oranges.handleInsertManyResponse([1], oranges),
+    [
+      { orangeId: 1, orangeType: 'valencia' },
+      { orangeType: 'navel' }
+    ],
+    'returns records with one updated (single ID response)'
+  );
+
+  st.deepEqual(
+    Oranges.handleInsertManyResponse([1, 2], oranges),
+    [
+      { orangeId: 1, orangeType: 'valencia' },
+      { orangeId: 2, orangeType: 'navel' }
+    ],
+    'returns updated records (multi ID response)'
+  );
+
+  st.end();
+});
+
+
+test('Mapper.defaultAttributes().prepareInsert()', st => {
+
+  const Defaults = Mapper.table('table').defaultAttributes({
+    default: 'default'
   });
 
-  t.test(
-    'Mapper.handleInsertManyResponse() - object response - ' +
-    'respects `attributeToColumn`',
-  t => {
+  st.queriesEqual(
+    Defaults.prepareInsert({ default: 'override' }).toQueryBuilder(),
+    `insert into "table" ("default") values ('override')`
+  );
 
-    const Oranges = Mapper.extend(CamelCase())
-      .table('oranges').idAttribute('orangeId');
+  st.queriesEqual(
+    Defaults.prepareInsert({}).toQueryBuilder(),
+    `insert into "table" ("default") values ('default')`
+  );
 
-    const oranges = [{ orangeType: 'valencia' }, { orangeType: 'navel'}];
-
-    const response = [
-      { orange_id: 1, orange_type: 'valencia', created_at: 'some_date' },
-      { orange_id: 2, orange_type: 'navel', created_at: 'some_time' }
-    ];
-
-    t.deepEqual(
-      Oranges.handleInsertManyResponse(response, oranges),
-      [
-        { orangeId: 1, orangeType: 'valencia', createdAt: 'some_date' },
-        { orangeId: 2, orangeType: 'navel', createdAt: 'some_time' }
-      ],
-      'returns updated records'
-    );
-
-    t.end();
+  const FnDefaults = Mapper.table('table').setState({
+    testState: 'state_value'
+  }).defaultAttributes({
+    default: 'default_value',
+    default_fn(attributes) {
+      return this.requireState('testState') + '/' + attributes.set;
+    }
   });
 
+  st.queriesEqual(
+    FnDefaults.prepareInsert({ set: 'set_value' }).toQueryBuilder(), `
+    insert into "table"
+      ("default", "default_fn", "set")
+    values
+      ('default_value', 'state_value/set_value', 'set_value')
+    `
+  );
 
-  t.test(
-    'Mapper.handleInsertManyResponse() - idResponse - ' +
-    'respects `attributeToColumn`',
-  st => {
+  st.end();
+});
 
-    const Oranges = Mapper.extend(CamelCase())
-      .table('oranges').idAttribute('orangeId');
+test('Mapper.strictAttributes().prepareInsert()', st => {
 
-    const oranges = [{ orangeType: 'valencia' }, { orangeType: 'navel'}];
-
-    st.deepEqual(
-      Oranges.handleInsertManyResponse([1], oranges),
-      [
-        { orangeId: 1, orangeType: 'valencia' },
-        { orangeType: 'navel' }
-      ],
-      'returns records with one updated (single ID response)'
-    );
-
-    st.deepEqual(
-      Oranges.handleInsertManyResponse([1, 2], oranges),
-      [
-        { orangeId: 1, orangeType: 'valencia' },
-        { orangeId: 2, orangeType: 'navel' }
-      ],
-      'returns updated records (multi ID response)'
-    );
-
-    st.end();
+  const Strict = Mapper.table('table').strictAttributes({
+    strict: 'strict_value'
   });
 
+  st.queriesEqual(
+    Strict.prepareInsert({ strict: 'overridden' }).toQueryBuilder(),
+    `insert into "table" ("strict") values ('strict_value')`
+  );
 
-  t.test('Mapper.defaultAttributes().prepareInsert()', st => {
+  st.queriesEqual(
+    Strict.prepareInsert({}).toQueryBuilder(),
+    `insert into "table" ("strict") values ('strict_value')`
+  );
 
-    const Defaults = Mapper.table('table').defaultAttributes({
-      default: 'default'
-    });
-
-    st.queriesEqual(
-      Defaults.prepareInsert({ default: 'override' }).toQueryBuilder(),
-      `insert into "table" ("default") values ('override')`
-    );
-
-    st.queriesEqual(
-      Defaults.prepareInsert({}).toQueryBuilder(),
-      `insert into "table" ("default") values ('default')`
-    );
-
-    const FnDefaults = Mapper.table('table').setState({
-      testState: 'state_value'
-    }).defaultAttributes({
-      default: 'default_value',
-      default_fn(attributes) {
-        return this.requireState('testState') + '/' + attributes.set;
-      }
-    });
-
-    st.queriesEqual(
-      FnDefaults.prepareInsert({ set: 'set_value' }).toQueryBuilder(), `
-      insert into "table"
-        ("default", "default_fn", "set")
-      values
-        ('default_value', 'state_value/set_value', 'set_value')
-      `
-    );
-
-    st.end();
+  const FnStrict = Mapper.table('table').setState({
+    testState: 'state_value'
+  }).strictAttributes({
+    strict: 'strict_value',
+    strict_fn(attributes) {
+      return this.requireState('testState') + '/' + attributes.strict;
+    }
   });
 
-  t.test('Mapper.strictAttributes().prepareInsert()', st => {
+  st.queriesEqual(
+    FnStrict.prepareInsert({ strict: 'set' }).toQueryBuilder(), `
+      insert into "table" ("strict", "strict_fn")
+      values ('strict_value', 'state_value/set')
+    `
+  );
 
-    const Strict = Mapper.table('table').strictAttributes({
-      strict: 'strict_value'
-    });
+  st.end();
+});
 
-    st.queriesEqual(
-      Strict.prepareInsert({ strict: 'overridden' }).toQueryBuilder(),
-      `insert into "table" ("strict") values ('strict_value')`
-    );
+test('Mapper.defaultAttributes().strictAttributes().prepareInsert()', st => {
 
-    st.queriesEqual(
-      Strict.prepareInsert({}).toQueryBuilder(),
-      `insert into "table" ("strict") values ('strict_value')`
-    );
-
-    const FnStrict = Mapper.table('table').setState({
-      testState: 'state_value'
-    }).strictAttributes({
-      strict: 'strict_value',
-      strict_fn(attributes) {
-        return this.requireState('testState') + '/' + attributes.strict;
-      }
-    });
-
-    st.queriesEqual(
-      FnStrict.prepareInsert({ strict: 'set' }).toQueryBuilder(), `
-        insert into "table" ("strict", "strict_fn")
-        values ('strict_value', 'state_value/set')
-      `
-    );
-
-    st.end();
+  const DefaultStrict = Mapper.table('table').defaultAttributes({
+    default: 'default_value'
+  }).strictAttributes({
+    strict_fn: attributes => attributes.default + '/' + attributes.set
   });
 
-  t.test('Mapper.defaultAttributes().strictAttributes().prepareInsert()', st => {
+  st.queriesEqual(
+    DefaultStrict.prepareInsert({
+      strict_fn: 'overridden', set: 'set_value'
+    }).toQueryBuilder(), `
+      insert into "table" ("default", "set", "strict_fn")
+      values ('default_value', 'set_value', 'default_value/set_value')
+    `,
+    'resolves strict with defaulted attributes'
+  );
 
-    const DefaultStrict = Mapper.table('table').defaultAttributes({
-      default: 'default_value'
-    }).strictAttributes({
-      strict_fn: attributes => attributes.default + '/' + attributes.set
-    });
+  st.end();
+});
 
-    st.queriesEqual(
-      DefaultStrict.prepareInsert({
-        strict_fn: 'overridden', set: 'set_value'
-      }).toQueryBuilder(), `
-        insert into "table" ("default", "set", "strict_fn")
-        values ('default_value', 'set_value', 'default_value/set_value')
-      `,
-      'resolves strict with defaulted attributes'
-    );
+test('Mapper.extend(CamelCase()).prepareInsert()', st => {
 
-    st.end();
-  });
+  const Shapes = Mapper.table('shapes').extend(CamelCase());
 
-  t.test(
-    'Mapper.extend(CamelCase()).prepareInsert() - respects `attributeToColumn`',
-  st => {
+  const records = [
+    { shapeType: 'square', sideCount: 4 },
+    { shapeType: 'triangle', sideCount: 3 }
+  ];
 
-    const Shapes = Mapper.table('shapes').extend(CamelCase());
+  st.queriesEqual(
+    Shapes.prepareInsert(records).toQueryBuilder(), `
+      insert into "shapes" ("shape_type", "side_count")
+      values ('square', 4), ('triangle', 3)
+    `,
+    'respects `attributeToColumn`'
+  );
 
-    const records = [
-      { shapeType: 'square', sideCount: 4 },
-      { shapeType: 'triangle', sideCount: 3 }
-    ];
-
-    st.queriesEqual(
-      Shapes.prepareInsert(records).toQueryBuilder(), `
-        insert into "shapes" ("shape_type", "side_count")
-        values ('square', 4), ('triangle', 3)
-      `
-    );
-
-    st.end();
-  });
-
+  st.end();
 });
